core/requirements.txt
bcrypt==4.0.1
mysql-connector-python==8.0.33
